Update 'ContactInquiry' model

+++ codecraft-web/Models/ContactInquiry.cs
@@ -6,12 +6,12 @@ namespace codecraft_web.Models
     public class ContactInquiry
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required, DataType(DataType.EmailAddress), MaxLength(320)]
         public string Email { get; set; }
         [Required]
         public string Message { get; set; }
         [DataType(DataType.DateTime)]
-        public DateTime CreatedAt { get; set; } // TIMESTAMP
+        public DateTime CreatedAt { get; set; }
     }
 }
